Add update tag category sample
Added a sample to update the tag category. This worked fine, but
required a tweak to the templates. There is no return value from this
API call which triggered an error of type Ok. Updated the api_client
files to include a check for return type _or_ a successful response.

Signed-off-by: J.R. Garcia <jrg@vmware.com>