Regenerate SDK for v0.2.0
There have been numerous changes to the SDK, but most of them should not
be apparent to consumers. The biggest change is that rather than having
each gem have it's own API client, configuration, etc. there is now a
runtime gem that contains all of that information. Another significant
change is having everything available to generate the SDK inside of the
.openapi folder. It contains the templates (which can have PRs made
against it), the configurations for the openapi-generator, and the OpenAPI
specs used to generate the gems.

Signed-off-by: J.R. Garcia <jrg@vmware.com>