Handle array query parameters correctly
Apparently, array query params were not being handled correctly. I am not
sure exactly when this started or why I distinctly remember testing it and
having it succeed, but apparently the way that Ethon handles collections
for array parameters is not correct for what we are expecting. Simply
setting the "params_encoding" to :multi takes care of this. Set it to
:multi by default as it will always be needed and consumers should not
have to pass this in themselves.

Fixes #53

Signed-off-by: J.R. Garcia <jrg@vmware.com>