Fix use of paths with preexisting query parameters
In several of the paths that get sent to the ApiClient, there are query
parameters that already exist. These were getting removed as they are
technically not part of the path, but were not being kept because they
were not in the list of query parameters. This lead to 404s on calls
made to these paths as they were malformed once the query parameters
were dropped.

Fixes #57

Signed-off-by: J.R. Garcia <jrg@vmware.com>