Fix subsequent authentication bug
In some cases (certain versions of vCenter from my testing),
authentication would work for the first call and fail on all subsequent
calls. This appears to be do to the ApiClient's handling of cookies. This
commit modifies the cookie handling code to set the Configuration's API
key rather than setting the header. This really is the better option and
continues to persist rather than failing on subsequent calls.

Fixes #62

Signed-off-by: J.R. Garcia <jrg@vmware.com>