Only ship the contents of the lib dir in the gem
This is what we use in a lot of our projects and it works really well to reduce total file count, which is particularly impacting on Windows systems were malmare scanners have to scan each of these files.

Here's the before / after compared to the 0.4.5. install I had on disk already that included the gem files in the pkg dir:

38.5 megs on disk -> 11.5
4331 files -> 1423

Signed-off-by: Tim Smith <tsmith84@gmail.com>