Update README to point to latest released version
Issue #71 rightly points out that the README points to a version that is
not yet released. This is due to the fact that the code on this branch
is not able to properly be built into gems as the some filenames have
gone over the maximum character limit.

This change points to the latest version on RubyGems.org to avoid any
confusion.

Signed-off-by: J.R. Garcia <jrg@vmware.com>